backend_fastapi/requirement.txt
pandas
pillow
matplotlib
numpy
fastapi
uvicorn
python-multipart
tensorflow==2.12.0